+++ subprojects/platform-native/src/main/java/org/gradle/nativeplatform/toolchain/internal/compilespec/CppCompileSpec.java
@@ -16,8 +16,11 @@
 
 package org.gradle.nativeplatform.toolchain.internal.compilespec;
 
+import org.gradle.nativeplatform.CppLanguageStandard;
 import org.gradle.nativeplatform.toolchain.internal.NativeCompileSpec;
 
 public interface CppCompileSpec extends NativeCompileSpec {
+    CppLanguageStandard getLanguageStandard();
 
+    void setLanguageStandard(CppLanguageStandard cppStandard);
 }
